tests: Move remote server billing tests into test_stripe_remote_server.
Give the self-hosted server billing flows their own module, parallel
to test_stripe_remote_realm, so that test_stripe is left with the
Cloud realm flows.  As before, fixtures are keyed by test function
name and need no regeneration; list the module for full-suite fixture
regeneration.
